Restructure the package to only define FFTW on 0.7
